refactor(evmlib): modularize `merkle_payment_vault` and update contract
